#23: move header to correct directory
