+++ h/jdcloud_signer/Credential.h
