+++ h/jdcloud_signer/http/Scheme.h
