+++ h/jdcloud_signer/util/StringUtils.h
