+++ h/jdcloud_signer/util/crypto/HashResult.h
