+++ h/jdcloud_signer/util/crypto/HashingUtils.h
