+++ h/jdcloud_signer/util/crypto/Outcome.h
