+++ h/jdcloud_signer/util/logging/LogMacros.h
