+++ h/jdcloud_signer/util/logging/Logging.h
